moved buttons needed auth to different components in user and post pages
Also orginazed types by importing them from the index file only
