+++ front-end/src/app/types/JobType.ts
@@ -1,7 +1,8 @@
 export type JobType = {
-    cleaning?: true;
-    walkingTheDog?: true;
-    cuttingGrass?: true;
-    movingHeavyObjects?: true;
-    plumbering?: true;
-  };
+  cleaning?: true;
+  walkingTheDog?: true;
+  cuttingGrass?: true;
+  movingHeavyObjects?: true;
+  plumbering?: true;
+  random?: true;
+};
