+++ front-end/src/app/types/PostType.ts
@@ -1,0 +1,12 @@
+import { JobType } from "./JobType";
+
+export type post = {
+  _id: string;
+  user: string;
+  picture: string;
+  pictures: string[];
+  title: string;
+  description: string;
+  price: number;
+  type: JobType;
+};
